backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-firestore==2.13.1
newspaper3k==0.2.8
requests==2.31.0
lxml==4.9.3